+++ lib/swiss_knife.dart
@@ -6,4 +6,5 @@ export 'src/date.dart';
 export 'src/collections.dart';
 export 'src/utils.dart';
 export 'src/data.dart';
+export 'src/resource.dart';